Make compound a package and split its classes in base, cast and condition modules
